Se estan aplicando buenas practicas de programacion corriguiendo sintaxis mal hechas en setup.py, GUI.py y se han cambiado funciones de setup a handler_funtions.py

+++ main.py
@@ -1,30 +1,28 @@
-from Modules.GUI import MiApp
 import tkinter as tk
 import base64
-from Modules.icon import img
 import os
 
-tmp = open("tmp.ico", "wb")
-tmp.write(base64.b64decode(img))
-tmp.close()
+from Modules.GUI import MiApp
+from Modules.icon import img
 
 def main():
-    # Crear la ventana principal
-    ventana_principal = tk.Tk()
-    ventana_principal.title("Servicio Gestion IoC Csirt")
-    ventana_principal.iconbitmap('tmp.ico')
+    # Crear y guardar ícono temporal
+    with open("tmp.ico", "wb") as tmp:
+        tmp.write(base64.b64decode(img))
 
-    # Crear una instancia de la clase FirewallBlockerApp para la primera página
-    app1 = MiApp(ventana_principal)
-    ventana_principal.protocol("WM_DELETE_WINDOW", app1.on_closing)
+    try:
+        ventana_principal = tk.Tk()
+        ventana_principal.title("Servicio Gestión IoC Csirt")
+        ventana_principal.iconbitmap("tmp.ico")
 
-    # Iniciar el ciclo de eventos
-    ventana_principal.mainloop()
+        app1 = MiApp(ventana_principal)
+        ventana_principal.protocol("WM_DELETE_WINDOW", app1.on_closing)
 
-    # Eliminacion de archivo ico
-    os.remove("tmp.ico")
+        ventana_principal.mainloop()
 
+    finally:
+        if os.path.exists("tmp.ico"):
+            os.remove("tmp.ico")
 
-# Ejecutar la función main() al final del script
 if __name__ == "__main__":
     main()
